basic navigation implemented with meet the dancers issue-#22
