+++ DancingQueens/package.json
@@ -10,13 +10,16 @@
   },
   "dependencies": {
     "@react-navigation/native": "^7.1.5",
+    "@react-navigation/native-stack": "^7.3.9",
     "@react-navigation/stack": "^7.2.9",
     "expo": "~52.0.42",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.8",
     "react-native-safe-area-context": "^5.3.0",
-    "react-native-screens": "^4.10.0"
+    "react-native-screens": "^4.10.0",
+    "react-native-gesture-handler": "~2.20.2",
+    "react-native-reanimated": "~3.16.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
